[kernel] Replace timer wakeup list with AVL tree + cached earliest pointer
Replace the O(N)-per-tick linked-list timer scan with an AVL tree
keyed by absolute expiry tick. A cached `earliest_timer` pointer
makes the common per-tick check O(1) with no IRQ-disable. Insert,
cancel, and expiry processing are O(log N) with bounded IRQ-off
time per expired timer.

Also widen task_set_wakeup_timer() to accept u64 ticks, which
removes the chunking complexity from kernel_sleep().

+++ kernel/poll.c
@@ -223,7 +223,7 @@ poll_wait_on_cond(struct pollfd *fds, nfds_t nfds, int timeout, int cond_cnt)
             break;
          }
 
-         task_set_wakeup_timer(curr,(u32)(timeout_ticks - elapsed));
+         task_set_wakeup_timer(curr, timeout_ticks - elapsed);
       }
 
       prepare_to_wait_on_multi_obj(waiter);
